+++ src/model/gamble-mode.ts
@@ -1,5 +1,5 @@
 import { GambleResult } from './gamble-result';
 
 export interface GambleMode {
-    winnings: (gamblingAmount: number) => GambleResult;
+    winnings: (gamblingAmount: number, jackpotEnabled: boolean) => GambleResult;
 }